# services/corporate-payments-api/openapi/corporate-payments-api.yml
openapi: 3.0.3
info:
  title: Commerzbank Corporate Payments API
  description: |
    Enables corporate and small business customers to submit payment orders and retrieve
    account transaction information via a secure API. Modelled after the Commerzbank
    Corporate Payments API (v1.6.0) available at developer.commerzbank.com.
  version: 1.6.0
  contact:
    name: Commerzbank Developer Portal
    url: https://developer.commerzbank.com

servers:
  - url: https://api.commerzbank.com/v1
    description: Production
  - url: https://api-sandbox.commerzbank.com/v1
    description: Sandbox

security:
  - BearerAuth: []

paths:
  /payments:
    post:
      summary: Submit a payment order
      description: |
        Initiates a SEPA credit transfer on behalf of the authenticated corporate client.
        Returns the payment ID and initial status upon acceptance.
      operationId: submitPaymentOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PaymentRequest'
            example:
              creditorName: Acme GmbH
              creditorIban: DE89370400440532013000
              creditorBic: COBADEFFXXX
              amount: "1500.00"
              currency: EUR
              remittanceInformation: Invoice INV-2024-001
      responses:
        '201':
          description: Payment order accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentCreatedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /payments/{paymentId}:
    get:
      summary: Get payment status
      description: Returns the current status and details of a previously submitted payment order.
      operationId: getPaymentStatus
      parameters:
        - name: paymentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Payment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentStatusResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Payment not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /accounts/{iban}/transactions:
    get:
      summary: Get account transactions
      description: Returns a list of booked transactions for the specified corporate account.
      operationId: getAccountTransactions
      parameters:
        - name: iban
          in: path
          required: true
          schema:
            type: string
            pattern: '^[A-Z]{2}[0-9]{2}[A-Z0-9]{1,30}$'
          example: DE40200400600100728700
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date
          description: Filter transactions from this date (inclusive)
        - name: dateTo
          in: query
          schema:
            type: string
            format: date
          description: Filter transactions up to this date (inclusive)
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 100
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Account not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: OAuth 2.0 access token from Commerzbank Authorization Server

  schemas:
    PaymentRequest:
      type: object
      required:
        - creditorName
        - creditorIban
        - amount
        - currency
      properties:
        creditorName:
          type: string
          description: Full name of the payment recipient
          example: Acme GmbH
        creditorIban:
          type: string
          pattern: '^[A-Z]{2}[0-9]{2}[A-Z0-9]{1,30}$'
          description: IBAN of the recipient account
          example: DE89370400440532013000
        creditorBic:
          type: string
          pattern: '^[A-Z]{6}[A-Z0-9]{2}([A-Z0-9]{3})?$'
          description: BIC/SWIFT code of the recipient's bank
          example: COBADEFFXXX
        amount:
          type: string
          pattern: '^\d+\.\d{2}$'
          description: Transfer amount as a decimal string
          example: "1500.00"
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
          description: ISO 4217 currency code
          example: EUR
        remittanceInformation:
          type: string
          maxLength: 140
          description: Payment reference / purpose
          example: Invoice INV-2024-001

    PaymentCreatedResponse:
      type: object
      required:
        - paymentId
        - status
        - timestamp
      properties:
        paymentId:
          type: string
          format: uuid
          description: Unique identifier for the submitted payment
        status:
          type: string
          enum: [RECEIVED, PENDING, ACSC, RJCT]
          description: |
            ISO 20022 payment status code:
            RECEIVED - Payment received by the bank
            PENDING - Payment is being processed
            ACSC - AcceptedSettlementCompleted
            RJCT - Rejected
        timestamp:
          type: string
          format: date-time
          description: Timestamp when the payment was received

    PaymentStatusResponse:
      type: object
      required:
        - paymentId
        - status
      properties:
        paymentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [RECEIVED, PENDING, ACSC, RJCT]
        creditorName:
          type: string
        creditorIban:
          type: string
        amount:
          type: string
        currency:
          type: string
        executionDate:
          type: string
          format: date

    Transaction:
      type: object
      required:
        - transactionId
        - bookingDate
        - amount
        - currency
      properties:
        transactionId:
          type: string
          description: Unique transaction identifier
        bookingDate:
          type: string
          format: date
          description: Date the transaction was booked
        valueDate:
          type: string
          format: date
          description: Value date of the transaction
        amount:
          type: string
          description: Transaction amount (negative for debits)
        currency:
          type: string
          pattern: '^[A-Z]{3}$'
        creditorName:
          type: string
          description: Name of the payment recipient (for debits)
        debtorName:
          type: string
          description: Name of the payment sender (for credits)
        remittanceInformation:
          type: string

    TransactionListResponse:
      type: object
      required:
        - transactions
        - totalCount
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        totalCount:
          type: integer
          description: Total number of transactions matching the query

    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Machine-readable error code
          example: UNAUTHORIZED
        message:
          type: string
          description: Human-readable error description
          example: Access token is missing or invalid